refactor: split
(cherry picked from commit 5683aeaefed71e91f13548a0014392bdf659d466)
