+++ frb_codegen/src/library/codegen/parser/mir/parser/function/real/mod.rs
@@ -32,10 +32,12 @@ use std::fmt::Debug;
 use syn::*;
 use MirSkipReason::IgnoreBecauseFunctionNotPub;
 use MirType::Primitive;
+use crate::codegen::parser::mir::parser::function::real::owner::OwnerInfoOrSkip;
 
 pub(crate) mod argument;
 pub(crate) mod output;
 mod transformer;
+mod owner;
 
 pub(crate) fn parse(
     src_fns: &[HirFlatFunction],
@@ -207,114 +209,6 @@ impl<'a, 'b> FunctionParser<'a, 'b> {
         }))
     }
 
-    fn parse_owner(
-        &mut self,
-        func: &HirFlatFunction,
-        context: &TypeParserParsingContext,
-        actual_method_dart_name: Option<String>,
-        attributes: &FrbAttributes,
-    ) -> anyhow::Result<OwnerInfoOrSkip> {
-        use MirSkipReason::*;
-        use OwnerInfoOrSkip::*;
-
-        match &func.owner {
-            HirFlatFunctionOwner::Function => Ok(Info(MirFuncOwnerInfo::Function)),
-            HirFlatFunctionOwner::StructOrEnum {
-                impl_ty,
-                trait_def_name,
-            } => {
-                let owner_ty = if let Some(x) = self.parse_method_owner_ty(impl_ty, context)? {
-                    x
-                } else {
-                    return Ok(Skip(IgnoreBecauseParseMethodOwnerTy));
-                };
-
-                let trait_def = if let Some(trait_def_name) = trait_def_name {
-                    if let Some(ans) = parse_type_trait(trait_def_name, self.type_parser) {
-                        Some(ans)
-                    } else {
-                        // If cannot find the trait, we directly skip the function currently
-                        return Ok(Skip(IgnoreBecauseParseOwnerCannotFindTrait));
-                    }
-                } else {
-                    None
-                };
-
-                if !is_allowed_owner(&owner_ty, attributes) {
-                    return Ok(Skip(IgnoreBecauseNotAllowedOwner));
-                }
-
-                self.parse_method_owner_inner(func, actual_method_dart_name, owner_ty, trait_def)
-            }
-            HirFlatFunctionOwner::TraitDef { trait_def_name } => {
-                let trait_def = MirTypeTraitDef {
-                    name: trait_def_name.to_owned(),
-                };
-
-                self.parse_method_owner_inner(
-                    func,
-                    actual_method_dart_name,
-                    MirType::TraitDef(trait_def.clone()),
-                    Some(trait_def),
-                )
-            }
-        }
-    }
-
-    fn parse_method_owner_inner(
-        &mut self,
-        func: &HirFlatFunction,
-        actual_method_dart_name: Option<String>,
-        owner_ty: MirType,
-        trait_def: Option<MirTypeTraitDef>,
-    ) -> anyhow::Result<OwnerInfoOrSkip> {
-        use OwnerInfoOrSkip::*;
-
-        let sig = func.item_fn.sig();
-        let mode = if matches!(sig.inputs.first(), Some(FnArg::Receiver(..))) {
-            MirFuncOwnerInfoMethodMode::Instance
-        } else {
-            MirFuncOwnerInfoMethodMode::Static
-        };
-
-        if owner_ty.should_ignore(self.type_parser) {
-            return Ok(Skip(IgnoreBecauseOwnerTyShouldIgnore));
-        }
-
-        let actual_method_name = sig.ident.to_string();
-
-        Ok(Info(MirFuncOwnerInfo::Method(MirFuncOwnerInfoMethod {
-            owner_ty,
-            actual_method_name,
-            actual_method_dart_name,
-            mode,
-            trait_def,
-        })))
-    }
-
-    fn parse_method_owner_ty(
-        &mut self,
-        impl_ty: &Type,
-        context: &TypeParserParsingContext,
-    ) -> anyhow::Result<Option<MirType>> {
-        let self_ty_path = if let Type::Path(self_ty_path) = impl_ty {
-            self_ty_path
-        } else {
-            return Ok(None);
-        };
-
-        // let owner_ty_name = external_impl::parse_name_or_original(
-        //     &(self_ty_path.path.segments.first().unwrap().ident).to_string(),
-        // )?;
-        let owner_ty_name = (self_ty_path.path.segments.first().unwrap().ident).to_string();
-        let syn_ty: Type = parse_str(&owner_ty_name)?;
-        Ok(Some(self.type_parser.parse_type(&syn_ty, context)?))
-    }
-}
-
-enum OwnerInfoOrSkip {
-    Info(MirFuncOwnerInfo),
-    Skip(MirSkipReason),
 }
 
 fn create_output_skip(func: &HirFlatFunction, reason: MirSkipReason) -> MirFuncOrSkip {
@@ -403,11 +297,6 @@ fn refine_namespace(owner: &MirFuncOwnerInfo) -> Option<Namespace> {
     } else {
         None
     }
-}
-
-fn is_allowed_owner(owner_ty: &MirType, attributes: &FrbAttributes) -> bool {
-    // if `#[frb(external)]`, then allow arbitrary type
-    attributes.external() || is_struct_or_enum_or_opaque_from_them(owner_ty)
 }
 
 pub(crate) fn is_struct_or_enum_or_opaque_from_them(ty: &MirType) -> bool {
